--- python_modules/dagster/dagster/tutorials/intro_tutorial/execution_context.yml
@@ -1,4 +1,0 @@
-context:
-  default:
-    config:
-      log_level: DEBUG
