+++ python_modules/dagster/dagster/tutorials/intro_tutorial/logging.yml
@@ -1,0 +1,3 @@
+logging:
+  console:
+    log_level: DEBUG
